+++ CHANGELOG.md
@@ -2,6 +2,13 @@ The v1.13.x release series is supported until **August 1 2027**.
 
 ## 1.13.0 (Unreleased)
 
+UPGRADE NOTES:
+
+- The "winrm" connection type for provisioners is no longer supported. ([#4012](https://github.com/opentofu/opentofu/pull/4012))
+
+    This connection type was deprecated in OpenTofu v1.12, and now removed in v1.13. Some of the upstream libraries OpenTofu was using to implement these features are no longer maintained, so it's not viable for us to offer this anymore.
+
+    [Modern Windows versions now support OpenSSH](https://learn.microsoft.com/en-us/windows-server/administration/openssh/openssh_install_firstuse), and so we suggest that anyone currently relying on WinRM plan to migrate to using SSH instead.
 
 ## Previous Releases