+++ CHANGELOG.md
@@ -1,4 +1,7 @@
-## 1.10.10 (unreleased)
+## 1.10.11 (unreleased)
+
+
+## 1.10.10
 
 BUG FIXES: